requirements.txt
busypie==0.5.1
http3==0.6.7
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.8
pymongo==4.3.3
grpcio-tools